refactor(api,simulators-api): fixed lint issues
