--- libs/datamodel/common/src/lib/core/scope.ts
@@ -1,9 +1,0 @@
-export interface Scope {
-    audience: string;
-    id: string;
-    description: string;
-}
-
-export type ScopeGroup = {[scope: string]: Scope};
-
-export type Scopes = {[id: string]: ScopeGroup};
